requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
